refactor(airc-cli): split monitor formatter modules
